fix(audit): correctifs P0/P1 audit securite + conformite facturation
Correctifs issus de l'audit complet du 2026-07-05 (voir docs/AUDIT-COMPLET-2026-07-05.md).

Securite:
- retrait des cles API reelles (Anthropic + FAL) en clair dans les .md
  (⚠️ historique git a purger + cles a revoquer cote fournisseurs)
- chat-marketing: rate-limit 15/min/IP + bornage messages (<=20, <=8000 car., roles valides)
- getClientIp: lecture x-forwarded-for/x-real-ip fonctionnelle sur Vercel
- bcrypt refresh token: cost 10 -> 12 (coherence + OWASP)

Conformite facturation:
- numerotation devis/factures delete-safe (max au lieu de count) + verrou
  consultatif Postgres partage entre les 2 chemins F- (plus de doublon ni collision)
- AVOIR: montants negatifs (normalisation idempotente des lignes)
- SOLDE: montantDeja borne au total TTC (net a payer >= 0)
- stats: arithmetique en Decimal (fin de l'accumulation flottante sur CA/marge)
- TVA: parseFloat au lieu de parseInt (taux 5,5% n'est plus tronque a 5%) [LignesEditor]

Donnees (migration 20260705_fk_setnull_token_unique):
- onDelete Cascade -> SetNull (Quote/Invoice/SignatureRequest) et Restrict
  (PaymentRequest): supprimer un dossier n'efface plus les pieces comptables
- token @unique sur Quote/Invoice (lookup e-facturation direct)
  (base prod verifiee: 0 doublon, tables vides -> migration sans risque)

Divers:
- stock: @Min(0) sur quantity/prix
- vercel.json (config active): bloc crons (relances 9h) retabli

Note: quotes.service.ts embarque aussi un petit garde-fou pre-existant
(reconversion devis deja facture) present dans le working tree avant l'audit.

+++ apps/web/app/(app)/facturation/components/LignesEditor.tsx
@@ -61,7 +61,7 @@ export const LignesEditor = React.memo(function LignesEditor({ lignes, onChange 
             <select
               className="rounded-lg border border-[#304035]/10 px-1 py-1.5 text-xs text-[#304035] bg-white focus:outline-none w-full"
               value={l.tva}
-              onChange={e => updateLigne(l.id, 'tva', parseInt(e.target.value))}
+              onChange={e => updateLigne(l.id, 'tva', parseFloat(e.target.value))}
             >
               {[0, 5.5, 10, 20].map(t => <option key={t} value={t}>{t}%</option>)}
             </select>
